updates
pull request

encapsulated the attack, jump and dash in the methods handle_attack, handle_jump, handle_dash, respectively, to help readability;

added "coyote time" which is split between "coyote edge" that gives a fraction of a second to jump after the character is off an edge and "coyote light feet" that allows the character to jump right before actually touching the ground to help with the responsiveness of the character control when jumping between platforms since the accuracy of thee machine check for a "character is on grounded" state doesn't match the human response;

added a method to increase the points the character gained;

added the basic pause mechanic;

added the inputs up(w), left(a), down(s), right(d), pause (esc), punch (enter/return), jump (space), guard(ctrl), change_camera_position(tab) and special(shift) under project settings;

added an autoloaded Instantiables script to hold all scenes that needs to be instantiated

added the folders textures, models (for raw models directly from blender for example), resources (inside script folder), text themes, materials, enemies (inside scenes folder) and stages (inside scenes folder) to help organise the project;

created new scene stages to test;

created a placeholder enemy to test how the interaction with badniks would be;

grouped the invisible walls and obstacles of the test map under a marker with the name "test map";

made basic hud life, special, avatar and points mechanics;

made the ability selection scene, still need to be implemented.

obs.
the handle_dash method was supposed to be use the _input method and a counter in _process to check double tap events and trigger the dash, but it didn't worked as intended, but at least it prevents the character from dashing when the player simply tried to press diagonally.