updates
serverJoin script is now autoloaded
add player method changed to configure player since it no longer spawns the player but configures it's multiplayer settings

restart buttons now works

pause menu added to the main menu scene for clarity

Instantiables script is a bit more clear now

global variables now have a current and select versions of areas, hubs and stages. the current versions are the ones that are instantiated in game and the selected versions are the ones selected in the main menu. this helps check which scene the player is on when loading new scenes, going back or restarting scenes.

pointer spawner now have a hud for the time remainging from ability select.