update
removed the commented lines in sonic script.
removed the punch_pressed variable that was not being used since the attack is stored in the attack_pressed variable.
moved the variable last_keycode, that stores the double tap, from the sonic script to the input script since it's only used there.
same for the bot script.